Added [QWERTY Keyboard](https://upload.wikimedia.org/wikipedia/commons/thumb/d/da/KB_United_States.svg/400px-KB_United_States.svg.png) Presses and started Mouse Movement

+++ kernel/src/lib/math.h
@@ -1,6 +1,6 @@
 #pragma once
 
 struct Point {
-	unsigned int x;
-	unsigned int y;
+	long x;
+	long y;
 };
